pyproject.toml
[tool.poetry]
name = "miniproj_lab_management"
version = "0.1.0"
description = "Mini Laboratory Management API"
authors = ["Soleil Lanuzo <soleillanuzo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.5"
djangorestframework = "3.13.1"
psycopg2 = "2.9.3"
django-enumfields = "2.1.1"
drf-yasg = "1.20.0"
django-filter = "22.1"
kafka-python = "2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"